Recovered from Eclipse.
git-svn-id: https://subversion.assembla.com/svn/cmosher@3158 884d43a4-ffbe-47e9-90ca-fd8c37d7eb32

+++ testjava/src/TestBeans.java
@@ -12,7 +12,7 @@ public class TestBeans
         }
         ed.setAsText("34");
         Integer i = (Integer)ed.getValue();
-        showInt(i);
+        showInt(i.intValue());
     }
 
     public static void showInt(int i)
